+++ src/test/java/com/smartsheet/api/internal/UserResourcesImplTest.java
@@ -30,9 +30,11 @@ import com.smartsheet.api.models.Result;
 import com.smartsheet.api.models.UserPlansResponse;
 import com.smartsheet.api.models.ListUsersWithFilters;
 import com.smartsheet.api.models.enums.ListUserInclusion;
+import com.smartsheet.api.models.enums.SeatType;
 import com.smartsheet.api.models.enums.UserInclusion;
 import com.smartsheet.api.models.enums.UserStatus;
 import org.assertj.core.util.Lists;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -215,6 +217,23 @@ class UserResourcesImplTest extends ResourcesImplBase {
         assertThat(updatedUser.getLicensedSheetCreator()).isTrue();
         assertThat(updatedUser.getId().longValue()).isEqualTo(8166691168380804L);
         assertThat(updatedUser.getStatus()).isEqualTo(UserStatus.ACTIVE);
+    }
+
+    @Test
+    void testUpgradeUser() throws IOException {
+        server.setResponseBody(new File("src/test/resources/upgradeUser.json"));
+        long userId = 123L;
+        long planId = 123L;
+        Assertions.assertDoesNotThrow(() -> userResources.upgradeUser(userId, planId, SeatType.UpgradeSeatType.MEMBER));
+    }
+
+    @Test
+    void testDowngradeUser() throws IOException {
+        server.setResponseBody(new File("src/test/resources/downgradeUser.json"));
+        long userId = 123L;
+        long planId = 123L;
+
+        Assertions.assertDoesNotThrow(() -> userResources.downgradeUser(userId, planId, SeatType.DowngradeSeatType.VIEWER));
     }
 
     @Test